requirements.txt
playwright
pytest
pytest-playwright
python-dotenv
Faker
pytest-asyncio
ruff
pytest-html